✨ Apply typescript to the project and fix a few errors

+++ src/resolvers/linkResolvers.ts
@@ -1,14 +1,21 @@
-// Resolvers
-export const linkResolvers = {
+import { ResolverMap } from "../types/interfaces.js";
+
+export const linkResolvers: ResolverMap = {
   Query: {
     info: () => `This is the API of a Hackernews Clone`,
     feed: (_, args, { linkService }) => linkService.getAllLinks(),
-    link: (_, { id }, { linkService }) => linkService.getLinkById(id),
+    link: (_, { id }, { linkService }) => {
+      if (!id) throw new Error('ID is required');
+      return linkService.getLinkById(id);
+    },
   },
   Mutation: {
     post: (_, { url, description }, { linkService, user }) => {
       if (!user) {
         throw new Error('Not authenticated');
+      }
+      if (!url || !description) {
+        throw new Error('URL and description are required');
       }
       return linkService.createLink(url, description, user.id);
     },
@@ -16,11 +23,17 @@ export const linkResolvers = {
       if (!user) {
         throw new Error('Not authenticated');
       }
+      if (!id) {
+        throw new Error('ID is required');
+      }
       return linkService.updateLink(id, { url, description });
     },
     deleteLink: (_, { id }, { linkService, user }) => {
       if (!user) {
         throw new Error('Not authenticated');
+      }
+      if (!id) {
+        throw new Error('ID is required');
       }
       return linkService.deleteLink(id);
     },
@@ -32,4 +45,4 @@ export const linkResolvers = {
       }).postedBy();
     }
   }
-};
+};
